Add initial draft of paper
Very rough, just a framework. We do not need any other sections,
but it is an option. (The paper must be very short, details can
go in the documentation.) The list of authors is completely made
up; probably it should reflect the list of contributors, and
include pivotal contributors from the past (with permission).

Towards https://github.com/equinor/scout/issues/1003.